So, now you can press F7 and the code will compile into an OpenGl application. And in this case it will be a raytracer, you can steer the camera with w,a,s,d or the arrows, and the mouse. (also left ctrl and space for up and down). The executable will be placed in a folder namned Blocks, which will be created in the project folder.
There was some problems with small blocks, so I thought that it is probably better to finish the big blocks, and then work towards the smaller ones if there is time... One of the reasons is that there existed a lot of ambiguity when the user wanted to modify a block or add some blocks. Bigger ones was a lot easier.